CN-129 done: home manager/logic

+++ Sources/Celler.App.Web/Game/Server/App/GameApplication.cs
@@ -73,7 +73,7 @@ namespace Celler.App.Web.Game.Server.App
 
         private void CreateTickTimer()
         {
-            var interval = TimeLogic.GetUpdateInterval()*1000;
+            var interval = (int)(TimeLogic.GetUpdateInterval()*1000);
             _tickTimer = new Timer( onTickTimer, null, 0, ( interval ) );
         }
